Added price-change and fixed held item issue
The price-change value in the config specifies how much the price will
increment/decrement when an item is sold. I also fixed an issue where if
the player tried to buy/sell/value with an empty hand, it would tell
them that the item held was not for sale (rather than telling them that
they weren't holding anything).